tests/wpt/web-platform-tests/navigation-api/scroll-restoration/manual-restoreScroll-after-dispatch.html

<!doctype html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<body>
<div style="height: 1000px; width: 1000px;"></div>
<div id="frag"></div>
<script>
promise_test(async t => {
  // Wait for after the load event so that the navigation doesn't get converted
  // into a replace navigation.
  await new Promise(resolve => window.onload = () => t.step_timeout(resolve, 0));
  assert_equals(window.scrollY, 0);
  await navigation.navigate("#frag").finished;
  assert_not_equals(window.scrollY, 0);

  let transitionWhile_resolve;
  let navigate_event;
  navigation.onnavigate = e => {
    navigate_event = e;
    e.transitionWhile(new Promise(r => transitionWhile_resolve = r), { scrollRestoration: "manual" });
  };
  let back_promises = navigation.back();
  await back_promises.committed;
  assert_not_equals(window.scrollY, 0);
  navigate_event.restoreScroll();
  assert_equals(window.scrollY, 0);
  transitionWhile_resolve();
  await back_promises.finished;
  assert_equals(window.scrollY, 0);
}, "scrollRestoration: restoreScroll() should work after a navigate event dispatch");
</script>
</body>
